Add DRep dependencies for test `2R`
